refactor: Make executor a service
