+++ tracecat/executor/models.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class ExecutorSyncInput(BaseModel):
+    origin: str
